api/requirements.txt
aerich==0.6.3
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
dictdiffer==0.9.0
fastapi==0.81.0
fastapi-utils==0.2.1
frozenlist==1.3.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
multidict==6.0.2
pydantic==1.9.1
pypika-tortoise==0.1.6
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomlkit==0.11.5
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
yarl==1.8.1
